+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.18.1 (Unreleased)
 
 - stdlib: `ContainsFunc` now allows its second argument to be null, to test whether the given collection contains any null elements.
+- stdlib: `MergeFunc` no longer panics if all of its arguments are null values of the same object type with at least one attribute.
 
 # 1.18.0 (February 23, 2026)